Move second number definition to numbers listener
Remove "let displayNumbers" and let "secondNumber"  from equalsButton
event listener and add the same code to an if statement within numbers
event listener.

As equalsButton is pressed when you're done writing operations and  the
calculator will not be able to perform more than one operation at the
same time.

Create an if statement whenever a number button is pressed
that if a first number has been defined before executes code
that defines the second number which can be used to perform operations
even if the equals button is not pressed.