Add evaluation of operations chained together
Add if statement (if secondNumber is defined and result is undefined)
to each operator button listener.

As each pair of numbers should be evaluated at a time, hence
that's why this if
statement is created, because it lets the calculator exactly do that.

This is done first by checking if secondNumber is defined (as it should
be defined, because that means the user wants to chain more operators)
and if result is not defined (it should not be defined because that
would mean the calculator has done an operation before this one) then
assigning the result of the operation of the previous two numbers and
displaying it on screen. After this, the second number is assigned to
undefined and result too (because we dont want to use them anymore and
could cause problems).