Fix variable name on function
Replace result with r on displayResult function.

As there's a global variable named result in the code, this name caused
problems on displayResult function because the statement "result =
undefined" should be calling the global variable instead of the local
variable. This bug is fixed by changing the name of the local variable
to "r".